fixes for some 'comparison between signed and unsigned integer expressions' warnings

+++ ugbase/lib_disc/spatial_disc/disc_util/fv1_geom.h
@@ -494,14 +494,14 @@ class FV1Geometry : public FVGeometryBase
 	/// return local coords for node ID
 		const MathVector<dim>& local_node_position(size_t nodeID) const
 		{
-			UG_ASSERT(nodeID < ref_elem_type::numCorners, "Invalid node id.");
+			UG_ASSERT(nodeID < (size_t) ref_elem_type::numCorners, "Invalid node id.");
 			return m_vvLocMid[0][nodeID];
 		}
 
 	/// return global coords for node ID
 		const MathVector<worldDim>& global_node_position(size_t nodeID) const
 		{
-			UG_ASSERT(nodeID < ref_elem_type::numCorners, "Invalid node id.");
+			UG_ASSERT(nodeID < (size_t) ref_elem_type::numCorners, "Invalid node id.");
 			return m_vvGloMid[0][nodeID];
 		}
